requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-experimental>=0.2.0
langgraph-sdk>=0.2.0
pydantic>=2.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.4.0
# MCP Dependencies
modelcontextprotocol>=0.1.0
langchain-mcp-adapters>=0.1.9
httpx>=0.24.0
